rename test and benchmark files, integrate parallel part in execute.c
