+++ raster/r.texture/testsuite/test_rtexture.py
