+++ raster/r.texture/testsuite/test_rtexture_parallel.py
